Implementation of CMB support in the runner
