+++ mapsims/example_config.cfg
@@ -28,6 +28,7 @@ so_pysm_models_version = "0.1"
 
 [ pysm_components ]
 
+    # Remove this if no default PySM components are needed
     pysm_components_string = "a1"
 
     [[ dust ]]
@@ -61,3 +62,18 @@ so_pysm_models_version = "0.1"
     curv = 0.
     nu_0 = 23.
     seed = 30001
+
+    [[ cmb ]]
+    class  =  mapsims.SOPrecomputedCMB
+    iteration_num = 0
+    nside = $output_nside
+    lensed = False
+    # At NERSC use:
+    # lensed = True
+    aberrated = False
+    has_polarization = True
+    cmb_set = 0
+    cmb_dir = "mapsims/data"
+    # At NERSC use:
+    # cmb_dir = "/global/project/projectdirs/sobs/v4_sims/mbs/cmb"
+    input_units = "uK_RJ"
